requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.1.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
python-dotenv==1.1.1
Werkzeug==3.1.3
Flask-WTF==1.2.2
email_validator==2.3.0
flask-migrate==4.1.0
flask-sqlalchemy==3.1.1
flask-bcrypt==1.0.1
flask-login==0.6.3
Pillow==11.1.0